#refactor: Moved the error folder to core
