+++ src/core/config/index.ts
@@ -1,5 +1,5 @@
 import JOOR_CONFIG from '@/types/config';
-import Jrror from '@/error';
+import Jrror from '@/core/error';
 import fs from 'node:fs';
 import path from 'node:path';
 import chalk from 'chalk';
