+++ src/core/internals/router/validators.ts
@@ -1,4 +1,4 @@
-import Jrror from '@/error';
+import Jrror from '@/core/error';
 import { ROUTE_HANDLER, ROUTE_PATH } from '@/types/route';
 
 function validateRoute(route: ROUTE_PATH) {
